pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ecological_emergent_behavior"
version = "0.0.0"
authors = [
  {name="Joseph Bejjani", email="jbejjani2022@gmail.com"},
  {name="Chase Van Amburg"},
  {name="Chengrui Wang"},
  {name="Chloe Su"},
  {name="Aaron Walsman", email="aaronwalsman@gmail.com"},
]
description = "Experiment code for 'The Emergence of Complex Behavior in Large-Scale Ecological Environments'"
readme = "README.md"
requires-python = ">=3.11,<3.13"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "jax",
  "chex",
  "wandb",
  "python-dotenv",
  "imageio[ffmpeg]",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["ecological_emergent_behavior*"]

[project.urls]
Homepage = "https://github.com/jbejjani2022/ecological-emergent-behavior"
Issues = "https://github.com/jbejjani2022/ecological-emergent-behavior/issues"
